+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - PR #2131 Chunk rows logic added to csv_writer
 - PR #2129 Add functions in the Java API to support nullable column filtering
 - PR #2165 made changes to get_dummies api for it to be available in MethodCache
+- PR #2176 Added column dtype support for datetime, int8, int16 to csv_writer
 
 ## Bug Fixes
 - PR #2086 Fixed quantile api behavior mismatch in series & dataframe